feat: build projects section.
include project cards that are dynamically generated using .map method on the src/data/project.js array of objects. this allows updates to be done on the projects through this file rather than in the jsx directly. ensure responsiveness of section by resizing cards. add hover effect to screenshots to take user to the live site.